hwmon: pm8xxx-adc: Remove wakelocks
The PM8xxx ADC interrupt is configured as a
wakeable interrupt. Therefore the system can suspend
and still have the end of conversion interrupt be
serviced.

Add a timeout in the wait_for_completion in case
the end of conversion event is not received. This
is to rule out any side affects if a missed interrupt
occurs.

Change-Id: I8fbf8328a538af3c0f5d233e2124daf5c2f675f8